acs: adds basic setup for standalone utility
This simply takes iw code and makes the iw survey dump
a stand alone utility which we will use for analysis of
the channel data. I decided to not use a screen scraper
on iw given that interpreters would only increase bloat
on embedded device. Hence this small utility application
approach.

The first step will be to use the dump data to look for
an ideal channel. mac80211_hswim only dumps data from
one channel, and ath*k drivers do it for all channels,
so for now ath*k drivers are required to actually make
some use of this.

Given that it was decided to not support independent
survey queries per frequency the goal will be to scrape
the entire data for each frequency query.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>