acs: change noise amplifier coefficient to 2
We change the noise amplifier coffecient  to 2 so we  now get:

                    (busy time - tx time)
    int_factor = ------------------------- *  2^(noise + min_noise)
                   (active time - tx time)

The coefficient of of 2 reflects the way power in "far-field" radiation
decreases as the square of distance from the antenna [1]. Thanks to
Felix Bitterli for the hint.

This reduction also means smaller results. If we want even smaller
values we can use log2() for the results but not sure what is best.

[1] http://en.wikipedia.org/wiki/Near_and_far_field

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>